motion file upload change to reflect api errors

+++ public/app/components/motion/motion.ctr.js
@@ -4,7 +4,7 @@
         .module('iserveu')
         .controller('MotionController', MotionController);
 
-    function MotionController($rootScope, $stateParams, $mdToast, $filter, $state, $location, $anchorScroll, $interval, motion,
+    function MotionController($rootScope, $stateParams, $mdToast, $filter, $state, $location, $anchorScroll, $interval, $timeout, motion,
     motionfile, UserbarService, ToastMessage) {
 
         var vm = this;
@@ -19,7 +19,9 @@
             title: null,
         }];
 
-        vm.goTo = function(id){
+        vm.goTo = goTo;
+
+        function goTo(id){
             $location.hash(id);
             $anchorScroll();
         }
@@ -49,6 +51,7 @@
             $mdToast.show(toast).then(function(response){
                 if(response == 'ok'){
                     vm.editMotion();
+                    cancelEdit();
                 }
             })
         }
@@ -123,7 +126,7 @@
         getMotion($stateParams.id);
 
         /**************************************** Motion Files Functions **************************************** */
-        // abstract this whole section and figure a way to avoid redundancy with these functions that are being used by edit motion ctrl 
+        // abstract this whole section and figure a way to avoid redundancy with these functions that are being used by create motion ctrl 
 
         vm.motion_files;
 
@@ -133,50 +136,109 @@
             file_id: ''
         }];
 
-        vm.formData = {};
+        vm.updated_motion = [{
+                file_category_name: "motionfiles",
+                title: '',
+                motion_id: '',
+                file_id: ''
+        }]
 
-        function getMotionFiles(id){    // unnecessary step 
+        vm.validate             = validate;
+        vm.formData             = {};
+        vm.errorFiles           = [];
+        vm.viewFiles            = [];
+        vm.deleteTempFiles      = [];
+        vm.tempResult           = [];
+        vm.updateTempMotionFile = [];
+
+        function getMotionFiles(id){  
             motionfile.getMotionFiles(id).then(function(result) {
                 vm.motion_files = result;
-
             })
         }
 
-        function motionFileLogic(){
-            if(vm.formData){uploadMotionFile($stateParams.id);}
-            if(vm.delete_motion_file){deleteMotionFiles();}
-            
-            updateMotionFiles();
+        function cancelEdit(){
+            angular.forEach(vm.tempValue, function(tempValue, key){
+                motionfile.deleteMotionFile(tempValue.motion_id, tempValue.file_id);
+            })
+            reset_variables();
         }
 
+        function motionFileLogic(){
+            deleteMotionFiles();
+            deleteTempMotionFiles();
+            updateMotionFiles();
+            updateTempMotionFiles();
+            reset_variables();
+        }
+
+        function upload (file) {
+            var fd = new FormData();
+            fd.append("file", file.file);
+            fd.append("file_category_name", "motionfiles");
+            fd.append("title", file.name);
+
+            motionfile.uploadMotionFile($stateParams.id, fd).then(function(tempResult){
+                tempResult.data["index"] = file.index;
+                vm.tempResult.push(tempResult.data);
+                deleteTempMotionFiles();
+            }, 
+
+            function(error){
+                vm.uploadError = true;
+                var error_message = error.data.message.substr(0, 20);
+                if(error_message == "The file is too big."){
+                    vm.errorFiles.push({file:file.file, error: error.data.message});
+                }
+            });
+        }
+
+        function validate(file, index){
+            file["index"] = vm.viewFiles.length;
+            if(!!{png:1,gif:1,jpg:1,jpeg:1,pdf:1}[file.getExtension()]){
+                vm.viewFiles.push(file);
+                upload(file);
+            }
+            else {
+                vm.uploadError = true;
+                vm.errorFiles.push({file:file, error: "File must be a png, jpeg, gif, jpg, or pdf."});
+            }
+        }
+
+        vm.changeTitleName = function(file){
+            vm.updateTempMotionFile.push(file);
+        }
 
         vm.updateMotionFile = function(file) {
             vm.updated_motion[file.file_id]= {
-                file_category_name: "motionfiles",
                 title: file.title,
                 motion_id: file.motion_id,
                 file_id: file.id
             }
         }
 
-        // title not working on post
         function updateMotionFiles(){
-            angular.forEach(vm.updated_motion, function(file, key) {
-                if(key != 0){
+            if(vm.updated_motion.length > 1){
+                angular.forEach(vm.updated_motion, function(file, key) {
                     motionfile.updateMotionFile(file, file.motion_id, file.file_id);
-                }
-            })
+                })
+            }
         }
 
-        vm.changeTitleName = function(index, name){
-            vm.formData[index].append("title", name);
+        function updateTempMotionFiles(){
+            if(vm.updateTempMotionFile.length > 0){
+                angular.forEach(vm.updateTempMotionFile, function(updateValue, updateKey){
+                    angular.forEach(vm.tempResult, function(tempValue, key) {
+                        if(updateValue.index == tempValue.index){
+                            delete vm.updateTempMotionFile[updateKey];
+                            motionfile.updateMotionFile({title: updateValue.name}, tempValue.motion_id, tempValue.file_id);
+                        }
+                    })
+                })
+            }
         }
 
-        vm.removeFile = function(index){
-            delete vm.formData[index];
-        }
-
-        vm.deleteMotionFile = function(bool, motion_id, file_id) {
+        vm.deleteOldMotionFile = function(bool, motion_id, file_id) {
             vm.delete_motion_file[file_id] = {
                 bool: !bool,
                 motion_id: motion_id,
@@ -184,29 +246,48 @@
             }
         }
 
-        vm.upload = function(flow) {
-            angular.forEach(flow.files, function(flowObject, index){
-                vm.formData[index] = new FormData();
-                vm.formData[index].append("file", flowObject.file);
-                vm.formData[index].append("file_category_name", "motionfiles");
-                vm.formData[index].append("title", flowObject.name);
-            })
-
-        }
-
-        function uploadMotionFile(id){
-            angular.forEach(vm.formData, function(value, key) {
-                motionfile.uploadMotionFile(id, value);
-            })
-        }
-
         function deleteMotionFiles(){
-            angular.forEach(vm.delete_motion_file, function(file, key) {
-                if(file.bool){
-                    motionfile.deleteMotionFile(file.motion_id, file.file_id);
-                }
-            })
+            if(vm.delete_motion_file.length > 0){
+                angular.forEach(vm.delete_motion_file, function(file, key) {
+                    if(file.bool){
+                        motionfile.deleteMotionFile(file.motion_id, file.file_id);
+                    }
+                })
+            }
         }
+
+        vm.deleteTempMotionFile = function(file){
+            vm.deleteTempFiles.push(file.file);
+        }
+
+        function deleteTempMotionFiles(){
+            if(vm.deleteTempFiles.length > 0){
+              angular.forEach(vm.deleteTempFiles, function(deleteValue, deleteKey){
+                    angular.forEach(vm.tempResult, function(tempValue, key) {
+                        if(deleteValue.name == tempValue.title){
+                            delete vm.deleteTempFiles[deleteKey];
+                            motionfile.deleteMotionFile(tempValue.motion_id, tempValue.file_id);
+                        }
+                    });
+                });  
+            }
+        }
+
+        function reset_variables(){
+            $timeout(function(){
+                vm.formData             = {};
+                vm.errorFiles           = [];
+                vm.viewFiles            = [];
+                vm.deleteTempFiles      = [];
+                vm.tempResult           = [];
+                vm.updateTempMotionFile = [];
+            }, 6000);
+        }
+
+        $rootScope.$on('createMotionAndAddAttachments', function(events, data){
+            vm.editMotionMode = true;
+            goTo('attachments');
+        })
 
     }
